+++ crates/hwpforge-smithy-hwp5/src/layout_hint_patch.rs
@@ -185,6 +185,9 @@ fn collect_flow_paragraph_layout_hints(
             | Hwp5Control::Line(_)
             | Hwp5Control::Rect(_)
             | Hwp5Control::Polygon(_)
+            | Hwp5Control::Ellipse(_)
+            | Hwp5Control::Arc(_)
+            | Hwp5Control::Curve(_)
             | Hwp5Control::OleObject(_)
             | Hwp5Control::Unknown { .. } => {}
         }
@@ -213,6 +216,9 @@ fn collect_scope_paragraph_layout_hints(paragraph: &Hwp5Paragraph, scope: &mut S
             | Hwp5Control::Line(_)
             | Hwp5Control::Rect(_)
             | Hwp5Control::Polygon(_)
+            | Hwp5Control::Ellipse(_)
+            | Hwp5Control::Arc(_)
+            | Hwp5Control::Curve(_)
             | Hwp5Control::OleObject(_)
             | Hwp5Control::Unknown { .. } => {}
         }